Moved interfaces to interfaces package
