+++ RPGManager_development/src/rpg_database/character_sheet/interfaces/CustomSetter.java
@@ -1,4 +1,6 @@
-package rpg_database.character_sheet;
+package rpg_database.character_sheet.interfaces;
+
+import rpg_database.character_sheet.CharacterSheet;
 
 public interface CustomSetter<DATA extends CustomSetter<DATA>> {
 	public Class<DATA> getImplementingClass();
